requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Flask-Limiter==3.5.0
python-socketio==5.9.0
python-engineio==4.8.0
python-dotenv==1.0.0
eventlet==0.33.3
PyJWT==2.8.0
